requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
kafka-python==2.0.2
psycopg2-binary==2.9.6
redis==4.5.5
sqlalchemy>=2.0.0
aiokafka>=0.8.0
asyncpg>=0.27.0